Extracted jquery-cookie.js from jquery.plugins.js
