+++ app/view/twig/_base/_page.twig
@@ -29,6 +29,7 @@
     'view/lib/jquery-tagcloud/jquery-tagcloud.min.js',
     'view/lib/jquery-hotkeys/jquery-hotkeys.min.js',
     'view/lib/jquery-watchchanges/jquery-watchchanges.min.js',
+    'view/lib/jquery-cookie-1.4.0/jquery-cookie.js',
     'view/js/jquery.plugins.js',
     'view/lib/jquery.fileupload-5.26/jquery.iframe-transport.js',
     'view/lib/jquery.fileupload-5.26/jquery.fileupload.js',
